feat: get_directory_all_entries
