+++ includes/directory.h
@@ -6,7 +6,7 @@
 /*   By: surayama <surayama@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/27 15:25:07 by surayama          #+#    #+#             */
-/*   Updated: 2025/11/27 16:04:51 by surayama         ###   ########.fr       */
+/*   Updated: 2025/11/27 16:08:52 by surayama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,8 @@
 # include <sys/stat.h>
 
 int	get_directory_entries(const char *path, bool include_hidden,
+		t_list **entries);
+int	get_directory_all_entries(const char *path, bool include_hidden,
 		t_list **entries);
 
 #endif
